Files containg line with string 'open' crashes
The keyword 'open' is used to initialize the transfer of a file. A file can be initialized and then initialized when another instance of 'open' is encountered within the same file, causing it to not open in Sublime-Text 3.  This issue can be easily recreated by adding the string 'open' (with no other characters) on a single line in a given file.

This change adds a check to make sure that the file was not already parsing data, solving this issue.